requirements.txt
alembic==1.15.2
aniso8601==10.0.0
attrs==25.3.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dotenv==0.9.9
exceptiongroup==1.2.2
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
PyJWT==2.10.1
pytest==8.3.5
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
rpds-py==0.23.1
SQLAlchemy==2.0.40
tomli==2.2.1
typing_extensions==4.13.0
Werkzeug==3.1.3
